moving new location APIs into a separate file
