+++ apps/teams-test-app/src/components/GeoLocationAPIs.tsx
@@ -1,0 +1,90 @@
+import { geoLocation, Location } from '@microsoft/teams-js';
+import React, { ReactElement } from 'react';
+
+import { ApiWithoutInput, ApiWithTextInput } from './utils';
+
+const CheckGeoLocationCapability = (): React.ReactElement =>
+  ApiWithoutInput({
+    name: 'checkLocationCapability',
+    title: 'Check geoLocation Capability',
+    onClick: async () => `geoLocation module ${geoLocation.isSupported() ? 'is' : 'is not'} supported`,
+  });
+
+const CheckGeoLocationMapCapability = (): React.ReactElement =>
+  ApiWithoutInput({
+    name: 'checkLocationMapCapability',
+    title: 'Check geoLocation Map Capability',
+    onClick: async () => `geoLocation module ${geoLocation.map.isSupported() ? 'is' : 'is not'} supported`,
+  });
+
+const GetCurrentLocation = (): React.ReactElement =>
+  ApiWithoutInput({
+    name: 'getCurrentLocation',
+    title: 'Get Current geoLocation',
+    onClick: async () => {
+      const result = await geoLocation.getCurrentLocation();
+      return JSON.stringify(result);
+    },
+  });
+
+const ChooseLocation = (): React.ReactElement =>
+  ApiWithoutInput({
+    name: 'map.chooseLocation',
+    title: 'Choose Location',
+    onClick: async () => {
+      const result = await geoLocation.map.chooseLocation();
+      return JSON.stringify(result);
+    },
+  });
+
+const ShowLocation = (): React.ReactElement =>
+  ApiWithTextInput<Location>({
+    name: 'map.showLocation',
+    title: 'Show Location',
+    onClick: {
+      validateInput: input => {
+        if (!input.latitude || !input.longitude) {
+          throw new Error('latitude and longitude are required');
+        }
+      },
+      submit: async locationProps => {
+        await geoLocation.map.showLocation(locationProps);
+        return 'Completed';
+      },
+    },
+  });
+
+const HasGeoLocationPermission = (): React.ReactElement =>
+  ApiWithoutInput({
+    name: 'HasGeoLocationPermission',
+    title: 'Has Permission',
+    onClick: async () => {
+      const result = await geoLocation.hasPermission();
+      return JSON.stringify(result);
+    },
+  });
+
+const RequestGeoLocationPermission = (): React.ReactElement =>
+  ApiWithoutInput({
+    name: 'RequestGeoLocationPermission',
+    title: 'Request Permission',
+    onClick: async () => {
+      const result = await geoLocation.requestPermission();
+      return JSON.stringify(result);
+    },
+  });
+
+const GeoLocationAPIs = (): ReactElement => (
+  <>
+    <h1>geoLocation</h1>
+    <GetCurrentLocation />
+    <ChooseLocation />
+    <ShowLocation />
+    <HasGeoLocationPermission />
+    <RequestGeoLocationPermission />
+    <CheckGeoLocationCapability />
+    <CheckGeoLocationMapCapability />
+  </>
+);
+
+export default GeoLocationAPIs;
