+++ packages/teams-js/src/public/index.ts
@@ -30,6 +30,7 @@ export { app } from './app';
 export { appInstallDialog } from './appInstallDialog';
 export { chat } from './chat';
 export { dialog } from './dialog';
+export { geoLocation } from './geoLocation';
 export { pages } from './pages';
 export { ChildAppWindow, IAppWindow, ParentAppWindow } from './appWindow';
 export { menus } from './menus';
