+++ packages/teams-js/test/public/location.spec.ts
@@ -1,8 +1,9 @@
 import { locationAPIsRequiredVersion } from '../../src/internal/constants';
 import { DOMMessageEvent } from '../../src/internal/interfaces';
 import { app } from '../../src/public/app';
-import { errorNotSupportedOnPlatform, FrameContexts } from '../../src/public/constants';
+import { FrameContexts } from '../../src/public/constants';
 import { ErrorCode, location, SdkError } from '../../src/public/index';
+import { Location } from '../../src/public/interfaces';
 import { _minRuntimeConfigToUninitialize } from '../../src/public/runtime';
 import { FramelessPostMocks } from '../framelessPostMocks';
 import { Utils } from '../utils';
@@ -14,13 +15,10 @@ describe('location', () => {
   const framelessPlatform = new FramelessPostMocks();
   const framedPlatform = new Utils();
   const minVersionForLocationAPIs = locationAPIsRequiredVersion;
-  const defaultLocationProps: location.LocationProps = { allowChooseLocation: false, showMap: false };
-  const defaultLocationPropsForChooseLocation: location.LocationProps = {
-    allowChooseLocation: true,
-    showMap: true,
-  };
-  const defaultLocation: location.Location = { latitude: 17, longitude: 17, accuracy: -1, timestamp: 100 };
+
+  const defaultLocation: Location = { latitude: 17, longitude: 17, accuracy: -1, timestamp: 100 };
   const originalDefaultPlatformVersion = '1.6.0';
+  const defaultLocationProps: location.LocationProps = { allowChooseLocation: false, showMap: false };
 
   beforeEach(() => {
     framelessPlatform.messages = [];
@@ -39,425 +37,8 @@ describe('location', () => {
   const emptyCallback = (): void => {
     return;
   };
-  const allowedContexts = [FrameContexts.content, FrameContexts.task];
 
-  describe('Testing getCurrentLocation API', () => {
-    it('should not allow getCurrentLocation calls before initialization', () => {
-      expect(() => location.getCurrentLocation()).toThrowError('The library has not yet been initialized');
-    });
-
-    Object.values(FrameContexts).forEach(context => {
-      if (allowedContexts.some(allowedContext => allowedContext === context)) {
-        it(`should throw error when getCurrentLocation is not supported in runtime config. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: {} });
-          try {
-            location.getCurrentLocation();
-          } catch (e) {
-            expect(e).toEqual(errorNotSupportedOnPlatform);
-          }
-        });
-
-        it(`getCurrentLocation calls with successful result. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-          const promise = location.getCurrentLocation();
-
-          const message = framelessPlatform.findMessageByFunc('location.getLocation');
-          expect(message).not.toBeNull();
-          expect(message.args.length).toBe(1);
-          expect(message.args[0]).toEqual(defaultLocationProps);
-
-          const callbackId = message.id;
-          framelessPlatform.respondToMessage({
-            data: {
-              id: callbackId,
-              args: [undefined, defaultLocation],
-            },
-          } as DOMMessageEvent);
-
-          await expect(promise).resolves.toBe(defaultLocation);
-        });
-
-        it(`getCurrentLocation calls with error. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-          const promise = location.getCurrentLocation();
-
-          const message = framelessPlatform.findMessageByFunc('location.getLocation');
-          expect(message).not.toBeNull();
-          expect(message.args.length).toBe(1);
-          expect(message.args[0]).toEqual(defaultLocationProps);
-
-          const callbackId = message.id;
-          framelessPlatform.respondToMessage({
-            data: {
-              id: callbackId,
-              args: [{ errorCode: ErrorCode.PERMISSION_DENIED }],
-            },
-          } as DOMMessageEvent);
-
-          await expect(promise).rejects.toEqual({ errorCode: ErrorCode.PERMISSION_DENIED });
-        });
-      } else {
-        it(`should not allow getCurrentLocation calls from the wrong context. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          expect(() => location.getCurrentLocation()).toThrowError(
-            `This call is only allowed in following contexts: ${JSON.stringify(
-              allowedContexts,
-            )}. Current context: "${context}".`,
-          );
-        });
-      }
-    });
-  });
-
-  describe('Testing HasPermisison API', () => {
-    it('should not allow hasPermission calls before initialization', () => {
-      return expect(() => location.hasPermission()).toThrowError('The library has not yet been initialized');
-    });
-
-    Object.values(FrameContexts).forEach(context => {
-      if (allowedContexts.some(allowedContext => allowedContext === context)) {
-        it(`should throw error when location is not supported in runtime config. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: {} });
-          expect.assertions(4);
-          try {
-            location.hasPermission();
-          } catch (e) {
-            expect(e).toEqual(errorNotSupportedOnPlatform);
-          }
-        });
-
-        it('hasPermission call in default version of platform support fails', async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          expect(() => location.hasPermission()).rejects.toEqual(errorNotSupportedOnPlatform);
-        });
-
-        it('hasPermission call with successful result', async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          const promise = location.hasPermission();
-
-          const message = framelessPlatform.findMessageByFunc('permissions.has');
-          expect(message).not.toBeNull();
-          expect(message.args.length).toBe(1);
-
-          const callbackId = message.id;
-          framelessPlatform.respondToMessage({
-            data: {
-              id: callbackId,
-              args: [undefined, true],
-            },
-          } as DOMMessageEvent);
-
-          await expect(promise).resolves.toBe(true);
-        });
-
-        it('HasPermission call with error', async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          const promise = location.hasPermission();
-
-          const message = framelessPlatform.findMessageByFunc('permissions.has');
-          expect(message).not.toBeNull();
-          expect(message.args.length).toBe(1);
-
-          const callbackId = message.id;
-          framelessPlatform.respondToMessage({
-            data: {
-              id: callbackId,
-              args: [{ errorCode: ErrorCode.INTERNAL_ERROR }],
-            },
-          } as DOMMessageEvent);
-
-          await expect(promise).rejects.toEqual({ errorCode: ErrorCode.INTERNAL_ERROR });
-        });
-      } else {
-        it(`should not allow hasPermission calls from the wrong context. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          expect(() => location.hasPermission()).toThrowError(
-            `This call is only allowed in following contexts: ${JSON.stringify(
-              allowedContexts,
-            )}. Current context: "${context}".`,
-          );
-        });
-      }
-    });
-  });
-
-  describe('Testing RequestPermisison API', () => {
-    Object.values(FrameContexts).forEach(context => {
-      if (allowedContexts.some(allowedContext => allowedContext === context)) {
-        it('should not allow requestPermission calls before initialization', () => {
-          expect(() => location.requestPermission()).toThrowError('The library has not yet been initialized');
-        });
-
-        it('requestPermission call in default version of platform support fails', async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(originalDefaultPlatformVersion);
-          expect(() => location.requestPermission()).rejects.toEqual(errorNotSupportedOnPlatform);
-        });
-
-        it(`should throw error when location is not supported in runtime config. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: {} });
-          expect.assertions(4);
-          try {
-            location.hasPermission();
-          } catch (e) {
-            expect(e).toEqual(errorNotSupportedOnPlatform);
-          }
-        });
-
-        it('requestPermission call with successful result', async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          const promise = location.requestPermission();
-
-          const message = framelessPlatform.findMessageByFunc('permissions.request');
-          expect(message).not.toBeNull();
-          expect(message.args.length).toBe(1);
-
-          const callbackId = message.id;
-          framelessPlatform.respondToMessage({
-            data: {
-              id: callbackId,
-              args: [undefined, true],
-            },
-          } as DOMMessageEvent);
-
-          await expect(promise).resolves.toBe(true);
-        });
-
-        it('requestPermission call with error', async () => {
-          await framelessPlatform.initializeWithContext(context);
-          framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-          const promise = location.requestPermission();
-
-          const message = framelessPlatform.findMessageByFunc('permissions.request');
-          expect(message).not.toBeNull();
-          expect(message.args.length).toBe(1);
-
-          const callbackId = message.id;
-          framelessPlatform.respondToMessage({
-            data: {
-              id: callbackId,
-              args: [{ errorCode: ErrorCode.INTERNAL_ERROR }],
-            },
-          } as DOMMessageEvent);
-
-          await expect(promise).rejects.toEqual({ errorCode: ErrorCode.INTERNAL_ERROR });
-        });
-      } else {
-        it(`should not allow requestPermission calls from the wrong context. context: ${context}`, async () => {
-          await framelessPlatform.initializeWithContext(context);
-          expect(() => location.requestPermission()).toThrowError(
-            `This call is only allowed in following contexts: ${JSON.stringify(
-              allowedContexts,
-            )}. Current context: "${context}".`,
-          );
-        });
-      }
-    });
-  });
-
-  describe('Testing location.map subcapability', () => {
-    describe('Testing location.map.showLocation API', () => {
-      it('should not allow showLocation calls before initialization', () => {
-        expect(() => location.map.showLocation(defaultLocation)).toThrowError(
-          'The library has not yet been initialized',
-        );
-      });
-
-      Object.values(FrameContexts).forEach(context => {
-        if (allowedContexts.some(allowedContext => allowedContext === context)) {
-          it(`should throw error when location is not supported in runtime config. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: {} });
-            try {
-              location.map.showLocation(defaultLocation);
-            } catch (e) {
-              expect(e).toEqual(errorNotSupportedOnPlatform);
-            }
-          });
-
-          it(`should throw error when location.map is not supported in runtime config. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-            try {
-              location.map.showLocation(defaultLocation);
-            } catch (e) {
-              expect(e).toEqual(errorNotSupportedOnPlatform);
-            }
-          });
-
-          it(`showLocation call in default version of platform support fails. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(FrameContexts.task);
-            framelessPlatform.setClientSupportedSDKVersion(originalDefaultPlatformVersion);
-            expect(location.map.showLocation(defaultLocation)).rejects.toEqual(errorNotSupportedOnPlatform);
-          });
-
-          it(`should not allow showLocation calls without props. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-            try {
-              location.map.showLocation(undefined);
-            } catch (e) {
-              expect(e).toEqual({ errorCode: ErrorCode.INVALID_ARGUMENTS });
-            }
-          });
-
-          it(`showLocation calls with successful result. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-            const promise = location.map.showLocation(defaultLocation);
-
-            const message = framelessPlatform.findMessageByFunc('location.showLocation');
-            expect(message).not.toBeNull();
-            expect(message.args.length).toBe(1);
-            expect(message.args[0]).toEqual(defaultLocation);
-
-            const callbackId = message.id;
-            framelessPlatform.respondToMessage({
-              data: {
-                id: callbackId,
-                args: [undefined, true],
-              },
-            } as DOMMessageEvent);
-
-            return expect(promise).resolves;
-          });
-
-          it(`showLocation calls with error context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-            const promise = location.map.showLocation(defaultLocation);
-
-            const message = framelessPlatform.findMessageByFunc('location.showLocation');
-            expect(message).not.toBeNull();
-            expect(message.args.length).toBe(1);
-            expect(message.args[0]).toEqual(defaultLocation);
-
-            const callbackId = message.id;
-            framelessPlatform.respondToMessage({
-              data: {
-                id: callbackId,
-                args: [{ errorCode: ErrorCode.PERMISSION_DENIED }],
-              },
-            } as DOMMessageEvent);
-
-            await expect(promise).rejects.toEqual({ errorCode: ErrorCode.PERMISSION_DENIED });
-          });
-        } else {
-          it(`should not allow showLocation calls from the wrong context. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            expect(() => location.map.showLocation(defaultLocation)).toThrowError(
-              `This call is only allowed in following contexts: ${JSON.stringify(
-                allowedContexts,
-              )}. Current context: "${context}".`,
-            );
-          });
-        }
-      });
-    });
-    describe('Testing location.map.chooseLocation API', () => {
-      it('should not allow chooseLocation calls before initialization', () => {
-        expect(() => location.map.chooseLocation()).toThrowError('The library has not yet been initialized');
-      });
-
-      Object.values(FrameContexts).forEach(context => {
-        if (allowedContexts.some(allowedContext => allowedContext === context)) {
-          it(`should throw error when location is not supported in runtime config. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: {} });
-            try {
-              location.map.chooseLocation();
-            } catch (e) {
-              expect(e).toEqual(errorNotSupportedOnPlatform);
-            }
-          });
-
-          it(`should throw error when location.map is not supported in runtime config. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-            try {
-              location.map.chooseLocation();
-            } catch (e) {
-              expect(e).toEqual(errorNotSupportedOnPlatform);
-            }
-          });
-
-          it(`chooseLocation call in default version of platform support fails. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(FrameContexts.task);
-            framelessPlatform.setClientSupportedSDKVersion(originalDefaultPlatformVersion);
-            expect(location.map.chooseLocation()).rejects.toEqual(errorNotSupportedOnPlatform);
-          });
-
-          it(`chooseLocation calls with successful result. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(FrameContexts.content);
-            framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-            const promise = location.map.chooseLocation();
-
-            const message = framelessPlatform.findMessageByFunc('location.getLocation');
-            expect(message).not.toBeNull();
-            expect(message.args.length).toBe(1);
-            expect(message.args[0]).toEqual(defaultLocationPropsForChooseLocation);
-
-            const callbackId = message.id;
-            framelessPlatform.respondToMessage({
-              data: {
-                id: callbackId,
-                args: [undefined, defaultLocation],
-              },
-            } as DOMMessageEvent);
-
-            await expect(promise).resolves.toBe(defaultLocation);
-          });
-
-          it(`chooseLocation calls with error context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(FrameContexts.content);
-            framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
-            framelessPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
-            const promise = location.map.chooseLocation();
-
-            const message = framelessPlatform.findMessageByFunc('location.getLocation');
-            expect(message).not.toBeNull();
-            expect(message.args.length).toBe(1);
-            expect(message.args[0]).toEqual(defaultLocationPropsForChooseLocation);
-
-            const callbackId = message.id;
-            framelessPlatform.respondToMessage({
-              data: {
-                id: callbackId,
-                args: [{ errorCode: ErrorCode.PERMISSION_DENIED }],
-              },
-            } as DOMMessageEvent);
-
-            await expect(promise).rejects.toEqual({ errorCode: ErrorCode.PERMISSION_DENIED });
-          });
-        } else {
-          it(`should not allow chooseLocation calls from the wrong context. context: ${context}`, async () => {
-            await framelessPlatform.initializeWithContext(context);
-            expect(() => location.map.chooseLocation()).toThrowError(
-              `This call is only allowed in following contexts: ${JSON.stringify(
-                allowedContexts,
-              )}. Current context: "${context}".`,
-            );
-          });
-        }
-      });
-    });
-  });
-
-  describe('testing deprecated location functions', () => {
+  describe('getLocation API', () => {
     it('should not allow getLocation calls before initialization', () => {
       expect(() => location.getLocation(defaultLocationProps, emptyCallback)).toThrowError(
         'The library has not yet been initialized',
@@ -466,6 +47,8 @@ describe('location', () => {
     it('getLocation call in default version of platform support fails', () => {
       framelessPlatform.initializeWithContext(FrameContexts.task).then(() => {
         framelessPlatform.setClientSupportedSDKVersion(originalDefaultPlatformVersion);
+        expect.assertions(4);
+
         try {
           location.getLocation(defaultLocationProps, emptyCallback);
         } catch (e) {
@@ -501,8 +84,11 @@ describe('location', () => {
     it('should not allow getLocation calls without props', () => {
       framedPlatform.initializeWithContext(FrameContexts.content).then(() => {
         framedPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
+        framedPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
+        expect.assertions(1);
+
         try {
-          location.getLocation(defaultLocationProps, emptyCallback);
+          location.getLocation(undefined, emptyCallback);
         } catch (e) {
           expect(e.errorCode).toBe(ErrorCode.INVALID_ARGUMENTS);
         }
@@ -547,7 +133,7 @@ describe('location', () => {
         framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
         framedPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
 
-        location.getLocation(defaultLocationProps, (error: SdkError, loc: location.Location) => {
+        location.getLocation(defaultLocationProps, (error: SdkError, loc: Location) => {
           expect(error).toBeFalsy();
           expect(loc).not.toBeNull();
           expect(loc.latitude).toBe(defaultLocation.latitude);
@@ -577,7 +163,7 @@ describe('location', () => {
         framelessPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
         framedPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
 
-        location.getLocation(defaultLocationProps, (error: SdkError, loc: location.Location) => {
+        location.getLocation(defaultLocationProps, (error: SdkError, loc: Location) => {
           expect(loc).toBeFalsy();
           expect(error.errorCode).toBe(ErrorCode.PERMISSION_DENIED);
           done();
@@ -597,7 +183,8 @@ describe('location', () => {
         } as DOMMessageEvent);
       });
     });
-
+  });
+  describe('Testing showLocation API', () => {
     it('should not allow showLocation calls before initialization', () => {
       expect(() => location.showLocation(defaultLocation, emptyCallback)).toThrowError(
         'The library has not yet been initialized',
@@ -607,6 +194,8 @@ describe('location', () => {
     it('showLocation call in default version of platform support fails', () => {
       framelessPlatform.initializeWithContext(FrameContexts.task).then(() => {
         framelessPlatform.setClientSupportedSDKVersion(originalDefaultPlatformVersion);
+        expect.assertions(4);
+
         try {
           location.showLocation(defaultLocation, emptyCallback);
         } catch (e) {
@@ -642,8 +231,11 @@ describe('location', () => {
     it('should not allow showLocation calls without props', () => {
       framedPlatform.initializeWithContext(FrameContexts.content).then(() => {
         framedPlatform.setClientSupportedSDKVersion(minVersionForLocationAPIs);
+        framedPlatform.setRuntimeConfig({ apiVersion: 1, supports: { location: {} } });
+        expect.assertions(1);
+
         try {
-          location.showLocation(defaultLocation, emptyCallback);
+          location.showLocation(undefined, emptyCallback);
         } catch (e) {
           expect(e.errorCode).toBe(ErrorCode.INVALID_ARGUMENTS);
         }
